4.21

+++ controllers/blogs.js
@@ -14,7 +14,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  
+  console.log('decooddu tokkeni', decodedToken)
   if (!decodedToken) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
@@ -61,8 +61,17 @@ blogsRouter.put('/:id', (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(404).json({ error: 'no permission to delete this' })
+  }
+
+  
 })
 
 module.exports = blogsRouter
